Changed name of node import toggle. Testing Github.
AS OF THIS VERSION:

-Marker importing does not work properly. They rely on node positions,
and I need to update the marker importing loop to work with the new node
system.

-The plugin still attempts to get vertex weights even if node importing
is disabled, resulting in an error. This can currently be avoided by
disabling vertex weights, and then disabling node importing in that
order, but this issue also leads into the next one.

-The UI is trash, and will be replaced soon. Instead of taking up real
estate full-time in the scene tab, the plugin will instead be invoked
through the File > Import drop-down. The user will select the file, and
a popup dialogue will follow for all the other options.

-Lots of stuff is spammed in the console. That's my ghetto way of
debugging. It'll all be cleaned out for release versions.